dayCycleScript created
Created a skeleton for the day cycle scripting. Need to add a timer of sorts.

Once we have other aspects of game I can update exactly what is needing to happen during the night.

Need to add when the user is ready to proceed to the next day.

Need to add attack checks.